requirements.txt
torch>=2.0.0
torchvision>=0.15.0

numpy>=1.15.3
pandas>=1.0.0
Pillow>=7.0.0
tqdm>=4.28.1

scikit-learn>=0.23.0
matplotlib>=3.0.0
opencv-python>=4.5.0
